+++ CHANGES.rst
@@ -11,7 +11,7 @@ Unreleased
 Version 2.3.3
 -------------
 
-Unreleased
+Released 2023-08-21
 
 -   Python 3.12 compatibility.
 -   Require Werkzeug >= 2.3.7.